requirements.txt
alabaster==0.7.12
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
beautifulsoup4==4.9.1
CacheControl==0.12.6
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudinary==1.20.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.2
django-ckeditor==6.0.0
django-cors-headers==3.4.0
django-easy-select2==1.5.7
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.9
django-htmlmin==0.11.0
django-js-asset==1.2.2
django-js-reverse==0.9.1
django-nine==0.2.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.16
drf-yasg==1.20.0
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
firebase-admin==4.4.0
google-api-core==1.23.0
google-api-python-client==1.12.8
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-cloud-core==1.4.3
google-cloud-firestore==2.0.1
google-cloud-storage==1.33.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
gunicorn==20.0.4
h11==0.11.0
html5lib==1.1
httplib2==0.18.1
idna==2.9
imagesize==1.2.0
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
oauth2client==4.1.3
packaging==20.4
proto-plus==1.11.0
protobuf==3.14.0
# psycopg2==2.8.5
psycopg2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-memcached==1.59
pytz==2020.1
requests==2.23.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
#six==1.14.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.25.9
uvicorn==0.12.1
webencodings==0.5.1
whitenoise==5.0.1
wrapt==1.12.1

boto3==1.17.97
botocore==1.20.97
django-storages==1.11.1
jmespath==0.10.0
python-dateutil==2.8.1
s3transfer==0.4.2
six==1.16.0